Move some common models to app "substance"
