+++ chemreg/substance/models.py
@@ -1,0 +1,27 @@
+from chemreg.common.models import ControlledVocabulary
+
+
+class Source(ControlledVocabulary):
+    """Controlled vocabulary for Sources
+
+    Attributes:
+        Name = String (Less than 50 character, url safe, unique, required field)
+        Label = String (Less than 100 characters, unique, required field)
+        Short Description = String (Less than 500 characters, required field)
+        Long Description = TEXT (required field)
+    """
+
+    pass
+
+
+class SubstanceType(ControlledVocabulary):
+    """Controlled vocabulary for Substances
+
+    Attributes:
+        Name = String (Less than 50 character, url safe, unique, required field)
+        Label = String (Less than 100 characters, unique, required field)
+        Short Description = String (Less than 500 characters, required field)
+        Long Description = TEXT (required field)
+    """
+
+    pass
