+++ chemreg/substance/tests/test_models.py
@@ -1,0 +1,37 @@
+from django.db import models
+
+from chemreg.substance.models import Source, SubstanceType
+
+
+def test_source_model():
+    """Tests the validity of the Source Model's attributes"""
+    name = Source._meta.get_field("name")
+    assert isinstance(name, models.SlugField)
+    assert name.max_length == 49
+    assert name.unique
+    label = Source._meta.get_field("label")
+    assert isinstance(label, models.CharField)
+    assert label.max_length == 99
+    assert label.unique
+    short_description = Source._meta.get_field("short_description")
+    assert isinstance(short_description, models.CharField)
+    assert short_description.max_length == 499
+    long_description = Source._meta.get_field("long_description")
+    assert isinstance(long_description, models.TextField)
+
+
+def test_substance_type():
+    """Tests the validity of the Substance Type Model's attributes"""
+    name = SubstanceType._meta.get_field("name")
+    assert isinstance(name, models.SlugField)
+    assert name.max_length == 49
+    assert name.unique
+    label = SubstanceType._meta.get_field("label")
+    assert isinstance(label, models.CharField)
+    assert label.max_length == 99
+    assert label.unique
+    short_description = SubstanceType._meta.get_field("short_description")
+    assert isinstance(short_description, models.CharField)
+    assert short_description.max_length == 499
+    long_description = SubstanceType._meta.get_field("long_description")
+    assert isinstance(long_description, models.TextField)
